requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgi-lifespan==2.1.0
asttokens==2.2.0
async-timeout==4.0.2
atlastk==0.13.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==4.1.0
black==23.1.0
bleach==5.0.1
celery==5.3.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
demjson==2.2.4
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
execnet==1.9.0
executing==1.2.0
fastapi==0.101.1
fastjsonschema==2.16.2
fqdn==1.5.1
frozenlist==1.3.3
h11==0.13.0
helium==3.0.9
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==5.1.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
iso8601==1.1.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.0.3
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kombu==5.3.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
motor==3.1.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
namedlist==1.8
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
openai==0.27.6
openpyxl==3.1.2
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.39
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pycrypto==2.6.1
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
Pygments==2.13.0
PyJWT==2.5.0
pymongo==4.2.0
PyMySQL==1.0.3
pyparsing==3.0.9
pypika-tortoise==0.1.6
pyrsistent==0.19.2
pytest==7.2.0
pytest-asyncio==0.21.1
pytest-xdist==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
redis==4.6.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
selenium==3.141.0
Send2Trash==1.8.0
sentry-sdk==1.9.8
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tortoise-orm==0.20.0
tqdm==4.64.1
traitlets==5.6.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
widgetsnbextension==4.0.4
xlrd==1.2.0
xlwt==1.3.0
yapf==0.32.0
yarl==1.9.2
zipp==3.11.0
